% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugParse.R
\name{drugParse}
\alias{drugParse}
\title{Parse TCGA clinical data drug information}
\usage{
drugParse(filenames, simplify = TRUE)
}
\arguments{
\item{filenames}{A vector contain the names of TCGA clinical XML files directorys and names (tcga_files_path/xml_file_names), which can easily retrieve by `list.files(path = ".", pattern = "*.xml",recursive = TRUE)`. The working directorys is the key, please check the names (which contain tcga_files_path/XML_file_names) is correct.}

\item{simplify}{logical, when simplify = TRUE return a tibble with patient ID information (bcr_patient_barcode, bcr_patient_uuid, tumor_tissue_site, gender), and a column contain
drug names. (it's a list column, thank to the tibble freature, and it contain the patient drug names has undergone).
When it's simplify = FALSE, it will return return a tibble with patient ID information (bcr_patient_barcode, bcr_patient_uuid, tumor_tissue_site, gender), and a column contain
full drug record.}
}
\value{
When simplify = TRUE, return a tibble with patients id and drug names list column. When simplify = FALSE, return a tibble contain full drugs record information.
}
\description{
`drugParse()` extract the drug record information from TCGA clinical XML files.

For each value in `filenames` `drugParse()` will extract the drug record information from clinical XML files downloaded from the TCGA database.
}
\author{
Fan Zhang
}
