% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow_upParse.R
\name{follow_upParse}
\alias{follow_upParse}
\title{Title Parse TCGA clinical follow up data from}
\usage{
follow_upParse(filenames, dir = ".", simplify = TRUE)
}
\arguments{
\item{filenames}{A vector contain the names of TCGA clinical XML files directorys and names (tcga_files_path/xml_file_names), which can easily retrieve by `list.files(path = ".", pattern = "*.xml",recursive = TRUE)`. The working directorys is the key, please check the names (which contain tcga_files_path/XML_file_names) is correct.}

\item{dir}{file directory contained the XML files, default is current directory "."}

\item{simplify}{should the output be simplified}
}
\description{
Title Parse TCGA clinical follow up data from
}
\author{
Fan Zhang <fanzhang95@outlook.com>
}
