% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_list.R
\name{join_list}
\alias{join_list}
\title{Title Combind a list of named vector into a data.frame by name}
\usage{
join_list(list)
}
\arguments{
\item{list}{a list of named vector, which each element has common name.}
}
\value{
a data frame
}
\description{
Title Combind a list of named vector into a data.frame by name
}
