% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageParse.R
\name{stageParse}
\alias{stageParse}
\title{Parse TCGA clinical data patients tumor stage information}
\usage{
stageParse(filenames, simplify = TRUE)
}
\arguments{
\item{filenames}{A vector contain the names of TCGA clinical XML files directorys and names (tcga_files_path/xml_file_names), which can easily retrieve by `list.files(path = ".", pattern = "*.xml",recursive = TRUE)`. The working directorys is the key, please check the names (which contain tcga_files_path/XML_file_names) is correct.}

\item{simplify}{When simplify = FALSE, it will return the full detail of patient basic information (bcr_patient_barcode, etc.), and tumor stage.
                When simplify = TRUE, it will return

When simplify = FALSE, the full information (including bcr_patient_barcode, bcr_patient_uuid, age_at_initial_pathologic_diagnosis, tumor_tissue_site, gender, race, stage_system_version, stage_clinical_stage, stage_pathologic_stage, stage_tnm_clinical_T,
stage_tnm_clinical_N, stage_tnm_clinical_M, stage_tnm_pathologic_T, stage_tnm_pathologic_M, stage_tnm_pathologic_N, stage_psa_value, stage_days_to_psa, stage_gleason_gleason_score, stage_gleason_primary_pattern, stage_gleason_secondary_pattern, stage_gleason_tertiary_pattern,
stage_ann_arbor_b_symptoms, stage_ann_arbor_extranodal_involvement, stage_serum_markers, stage_igcccg_stage, stage_masaoka_stage)

When simplify = TRUE, it will only return the key tumor stage information (bcr_patient_barcode, bcr_patient_uuid, age_at_initial_pathologic_diagnosis, tumor_tissue_site, gender, race, stage_system_version, stage_pathologic_stage, stage_tnm_pathologic_T, stage_tnm_pathologic_M, stage_tnm_pathologic_N)}
}
\value{
A tibble with every TCGA clinical xml files patients basic information (barcode, gender, etc.) and tumor stage information.
}
\description{
Parse TCGA clinical data patients tumor stage information
}
\author{
Fan Zhang
}
