% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumor_eventParse.R
\name{tumor_eventParse}
\alias{tumor_eventParse}
\title{Title Parse TCGA clinical tumor event data from XML files}
\usage{
tumor_eventParse(filenames, simplify = TRUE)
}
\arguments{
\item{filenames}{A vector contain the names of TCGA clinical XML files directorys and names (tcga_files_path/xml_file_names), which can easily retrieve by `list.files(path = ".", pattern = "*.xml",recursive = TRUE)`. The working directorys is the key, please check the names (which contain tcga_files_path/XML_file_names) is correct.}

\item{simplify}{should return be simplified ?}
}
\value{
a tibble with tumor event data
}
\description{
Title Parse TCGA clinical tumor event data from XML files
}
\author{
Fan Zhang
}
