% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitalParse.R
\name{vitalParse}
\alias{vitalParse}
\title{Parse TCGA clinical vital data from XML to OS and PFS for COX analysis}
\usage{
vitalParse(list, simplify = TRUE)
}
\arguments{
\item{list}{A vector names of TCGA clinical XML files directorys and names (tcga_files_path/xml_file_names), which can easily retrieve by `list.files(path = ".", pattern = "*.xml",recursive = TRUE)`. The working directorys is the key, please check the names (which contain tcga_files_path/XML_file_names) is correct.}

\item{simplify}{logical, when it's FLASE return TCGA vital information (including bcr_patient_barcode, vital_status, days_to_death, followupInfor_days_to_death,
days_to_last_followup, followupInfor_vital_status, followupInfor_days_to_last_followup).
when it's TURE, it return only (bcr_patient_barcode, days_to_last_followup, vital_status,days_to_new_tumor_event, tumor_event_status, tumor_event_free_days)
which is latest vital data (because the latest follow up vital data was integrated if avaliable) ready for survival analysis.}
}
\value{
If simplify=TRUE a data frame ready for survival analysis of OS and PFS, with vital information and new tumor event free times, where the new follow up vital data is integrated. As simplify=FALSE, a data frame contain all vital, follow up times and tumor event days will return.
}
\description{
`vitalParse()` integate TCGA clinical data from XML including the old vital data and updated latest follow up vital data. And the minimum days to new tumor event after initial treatment.

For each value in `list` (in fact, it's a vector) `vitalParse()` extract the vital data from clinical XML files downloaded from the TCGA database
including vital_status and days to last follow up (OS), the latest follow_up vital data, which is updataed vital data and it can be easily omited. Meanwhile the days to new tumor event (PFS) is the earliest days to tumor event, after_initial_treatment.
So we can use this OS, PFS to perform COX analysis. When simplify=TRUE,if the patient without new tumor event, it's pregression free time will be last follow up time.
}
\examples{
#list <- list.files(".","*.xml",recursive = TRUE)
#vitalParse(list,simplify=TRUE)
}
\author{
Fan Zhang
}
