% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageParse.R
\name{stageParse}
\alias{stageParse}
\title{Parse TCGA clinical data patients stage information}
\usage{
stageParse(filenames, simplify = TRUE)
}
\arguments{
\item{filenames}{A vector contain the names of TCGA clinical XML files directorys and names (tcga_files_path/xml_file_names), which can easily retrieve by `list.files(path = ".", pattern = "*.xml",recursive = TRUE)`. The working directorys is the key, please check the names (which contain tcga_files_path/XML_file_names) is correct.}

\item{simplify}{should return be simplified ?}
}
\description{
Parse TCGA clinical data patients stage information
}
\author{
Fan Zhang
}
