% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumor_eventParse.R
\name{tumor_eventParse}
\alias{tumor_eventParse}
\title{Title Parse TCGA clinical tumor event data from XML files}
\usage{
tumor_eventParse(filenames, dir = ".", simplify = TRUE)
}
\arguments{
\item{filenames}{A vector contain the names of TCGA clinical XML files directorys and names (tcga_files_path/xml_file_names), which can easily retrieve by `list.files(path = ".", pattern = "*.xml",recursive = TRUE)`. The working directorys is the key, please check the names (which contain tcga_files_path/XML_file_names) is correct.}

\item{dir}{file directory contained the XML files, default is current directory "."

tumor_eventParse() will extract the TCGA xml files, as you frome the follow up record of XML files.
Some patient may have multiple record of tumor event, others have none or

When simplify = TRUE, the returned tibble will contain column as following (bcr_patient_barcode, bcr_patient_uuid, age_at_initial_pathologic_diagnosis, tumor_tissue_site, gender, stage_event_system_version, stage_pathologic_stage, new_tumor_event_list).
last 3 column (new_tumor_neoplasm_event_type, new_tumor_occurrence_anatomic_site, new_neoplasm_event_type), is a list in a tibble column, due to one patient often have several new tumor metastasis event.

When simplify = FALSE, the returned tibble will contain bcr_patient_barcode, bcr_patient_uuid, age_at_initial_pathologic_diagnosis, tumor_tissue_site, gender, stage_event_system_version, stage_pathologic_stage, new_tumor_event_list.
The last one new_tumor_event_list is the full tumor event data, store in the tibble column, and each element is a table}

\item{simplify}{When simplify = TRUE, it will return the patients basic ID information, new_tumor_neoplasm_event_type, new_tumor_occurrence_anatomic_site, new_neoplasm_event_type.
When simplify = FALSE, it will return the patients basic ID information, and a tibble column, new_tumor_event_list, contain all the tumor event of the patient.}
}
\value{
a tibble with tumor event data
}
\description{
Title Parse TCGA clinical tumor event data from XML files
}
\seealso{
the vitalParse function for extract the vital data, also have the days to new tumor event, which can be used for PFS analysis.
 \code{link{vitalParse}}
}
\author{
Fan Zhang
}
