% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitalParse.R
\name{vitalParse}
\alias{vitalParse}
\title{Parse TCGA clinical vital data from XML for COX analysis}
\usage{
vitalParse(filenames, dir = ".")
}
\arguments{
\item{filenames}{A vector contain the names of TCGA clinical XML files directorys and names (tcga_files_path/xml_file_names), which can easily retrieve by `list.files(path = ".", pattern = "*.xml",recursive = TRUE)`. The working directorys is the key, please check the names (which contain tcga_files_path/XML_file_names) is correct.}

\item{dir}{file directory contained the XML files, default is current directory "."}
}
\description{
`vitalParse()` integate TCGA clinical data from XML including the old vital data and updated latest follow up vital data. And the minimum days to new tumor event after initial treatment.

For each value in `list` (in fact, it's a vector) `vitalParse()` extract the vital data from clinical XML files downloaded from the TCGA database
including vital_status and days to last follow up (OS), the latest follow_up vital data, which is updataed vital data and it can be easily omited. Meanwhile the days to new tumor event (PFS) is the earliest days to tumor event, after_initial_treatment.
So we can use this OS, PFS to perform COX analysis. return data including bcr_patient_barcode, tumor_tissue_site, vital_status, days_to_death, days_to_last_followup, new_tumor_event_after_initial_treatment, days_to_new_tumor_event_after_initial_treatment, days_tumor_free_till_last_followup
}
\author{
Fan Zhang <fanzhang95@outlook.com>
}
